backend/requirements.txt
pydantic-settings>=2.0.0
pydantic[email]
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
redis
celery
boto3
python-multipart
fastapi-limiter